KP-7440 Create conftest to share common fixtures
