+++ tests/conftest.py
@@ -1,0 +1,18 @@
+import pytest
+
+@pytest.fixture(autouse=True)
+def prevent_online_http_requests(monkeypatch):
+    """
+    Patch urlopen so that all non-patched requests raise an error.
+    """
+
+    def urlopen_error(self, method, url, *args, **kwargs):
+        raise RuntimeError(
+            f"Requests are not allowed in tests, but a test attempted a "
+            f"{method} request to {self.scheme}://{self.host}{url}"
+        )
+
+    monkeypatch.setattr(
+        "urllib3.connectionpool.HTTPConnectionPool.urlopen", urlopen_error
+    )
+
