+++ tests/test_metax_api.py
@@ -4,22 +4,6 @@ import requests_mock
 import metax_api
 import logging
 
-@pytest.fixture(autouse=True)
-def prevent_online_http_requests(monkeypatch):
-    """
-    Patch urlopen so that all non-patched requests raise an error.
-    """
-
-    def urlopen_error(self, method, url, *args, **kwargs):
-        raise RuntimeError(
-            f"Requests are not allowed in tests, but a test attempted a "
-            f"{method} request to {self.scheme}://{self.host}{url}"
-        )
-
-    monkeypatch.setattr(
-        "urllib3.connectionpool.HTTPConnectionPool.urlopen", urlopen_error
-    )
-
 @pytest.fixture
 def metax_base_url():
     return "https://metax-service.fd-staging.csc.fi/v3"
